backend/requirements.txt
# Django and REST framework
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9

# Blockchain
web3==6.15.1
eth-account==0.10.0
eth-typing==3.5.2
eth-utils==2.3.1

# Machine Learning
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3

# IoT and Messaging
paho-mqtt==1.6.1
pyserial==3.5
RPi.GPIO==0.7.1

# Chatbot
rasa==3.6.15
transformers==4.37.2
tensorflow==2.15.0

# Cloud Services
firebase-admin==6.4.0
boto3==1.34.34
django-storages==1.14.2
twilio==8.12.0
phonenumbers==8.13.30

# Development Tools
python-dotenv==1.0.1
pytest==7.4.4
pytest-django==4.7.0
black==24.1.1
isort==5.13.2
flake8==7.0.0

# Documentation
drf-yasg==1.21.7

# Logistics & GIS related
googlemaps==4.10.0
geopy==2.4.1
ortools==9.8.3296

# Training module related
django-embed-video==1.4.9
django-modeltranslation==0.18.11  # For content translations 